fix: remove MR check
