+++ tests/model_serving/model_server/conftest.py
@@ -4,6 +4,7 @@ import pytest
 import yaml
 from _pytest.fixtures import FixtureRequest
 from kubernetes.dynamic import DynamicClient
+from ocp_resources.cluster_service_version import ClusterServiceVersion
 from ocp_resources.config_map import ConfigMap
 from ocp_resources.inference_service import InferenceService
 from ocp_resources.namespace import Namespace
@@ -13,6 +14,8 @@ from ocp_resources.service_account import ServiceAccount
 from ocp_resources.serving_runtime import ServingRuntime
 from ocp_resources.storage_class import StorageClass
 from ocp_utilities.monitoring import Prometheus
+from pytest_testconfig import config as py_config
+from simple_logger.logger import get_logger
 
 from utilities.constants import (
     KServeDeploymentType,
@@ -34,6 +37,9 @@ from utilities.infra import (
     update_configmap_data,
 )
 from utilities.serving_runtime import ServingRuntimeFromTemplate
+
+
+LOGGER = get_logger(name=__name__)
 
 
 @pytest.fixture(scope="class")
@@ -563,3 +569,32 @@ def unprivileged_s3_caikit_serverless_inference_service(
         storage_path=request.param["model-dir"],
     ) as isvc:
         yield isvc
+
+
+@pytest.fixture(scope="session")
+def fail_if_missing_dependent_operators(admin_client: DynamicClient) -> None:
+    missing_operators: list[str] = []
+    csvs = list(
+        ClusterServiceVersion.get(
+            dyn_client=admin_client,
+            namespace=py_config["applications_namespace"],
+        )
+    )
+
+    for operator_name in py_config.get("dependent_operators", []).split(","):
+        LOGGER.info(f"Verifying if {operator_name} is installed")
+        for csv in csvs:
+            if csv.name.startswith(operator_name):
+                if csv.status == csv.Status.SUCCEEDED:
+                    break
+
+                else:
+                    missing_operators.append(
+                        f"Operator {operator_name} is installed but CSV is not in {csv.Status.SUCCEEDED} state"
+                    )
+
+        else:
+            missing_operators.append(f"{operator_name} is not installed")
+
+    if missing_operators:
+        pytest.fail(str(missing_operators))
